+++ torchbearer/cite.py
@@ -1,6 +1,6 @@
 def cite(bibtex):
-    """A decorator which adds a reference to the docstring of the given object. The docstring must contain ':bib:' which
-    is then replaced with the given bibtex string at runtime.
+    """A decorator which adds a reference to the **Google style** docstring of the given object. The ``Args:`` or
+    ``Returns:`` line is then prepended with the given bibtex string at runtime. Otherwise, the last line is used.
 
     Args:
         bibtex (str): The bibtex string to insert
@@ -8,10 +8,25 @@ def cite(bibtex):
     Returns:
         The decorator
     """
-    to_insert = '::\n\n' + ' '*8
-    to_insert += bibtex.strip().replace('\n', '\n' + ' '*8).rstrip()
-
     def decorator(inner):
-        inner.__doc__ = inner.__doc__.replace(':bib:', to_insert)
+        doc = inner.__doc__.split('\n')
+        i = 0
+        s = 0
+        for line in doc:
+            sline = line.strip()
+            if sline == 'Args:' or sline == 'Returns:':
+                for char in line:
+                    if char == ' ':
+                        s += 1
+                break
+            i += 1
+
+        spaces = ' ' * (s + 4)
+        to_insert = ' ' * s + '::\n\n' + spaces
+        to_insert += bibtex.strip().replace('\n', '\n' + spaces).rstrip()
+
+        doc.insert(i, '')
+        doc.insert(i, to_insert)
+        inner.__doc__ = '\n'.join(doc)
         return inner
     return decorator